// types/SelectedSpots.ts
type SelectedSpots = {
  spot8a: boolean;
  spot8b: boolean;
  spot8c: boolean;
  spot8d: boolean;
  spot8e: boolean;
  spot8f: boolean;
  spot8g: boolean;
  spot8h: boolean;
  spot7a: boolean;
  spot7b: boolean;
  spot7c: boolean;
  spot7d: boolean;
  spot7e: boolean;
  spot7f: boolean;
  spot7g: boolean;
  spot7h: boolean;
  spot6a: boolean;
  spot6b: boolean;
  spot6c: boolean;
  spot6d: boolean;
  spot6e: boolean;
  spot6f: boolean;
  spot6g: boolean;
  spot6h: boolean;
  spot5a: boolean;
  spot5b: boolean;
  spot5c: boolean;
  spot5d: boolean;
  spot5e: boolean;
  spot5f: boolean;
  spot5g: boolean;
  spot5h: boolean;
  spot4a: boolean;
  spot4b: boolean;
  spot4c: boolean;
  spot4d: boolean;
  spot4e: boolean;
  spot4f: boolean;
  spot4g: boolean;
  spot4h: boolean;
  spot3a: boolean;
  spot3b: boolean;
  spot3c: boolean;
  spot3d: boolean;
  spot3e: boolean;
  spot3f: boolean;
  spot3g: boolean;
  spot3h: boolean;
  spot2a: boolean;
  spot2b: boolean;
  spot2c: boolean;
  spot2d: boolean;
  spot2e: boolean;
  spot2f: boolean;
  spot2g: boolean;
  spot2h: boolean;
  spot1a: boolean;
  spot1b: boolean;
  spot1c: boolean;
  spot1d: boolean;
  spot1e: boolean;
  spot1f: boolean;
  spot1g: boolean;
  spot1h: boolean;
};

export default SelectedSpots;